Change required error message
Because the validation doesn't actually check that the name matches
anything, it only checks to make sure the box is filled out. It may not
be an especially useful validation, but we can't generically check that
the name matches what's in the form data because we don't know what the
data structure looks like for any given form.

What we _can_ do, however, is provide a validator function creator that
accepts the first and last names and returns the validator function.